backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai>=1.50.0
groq>=0.11.0
python-dotenv==1.0.0
pydub==0.25.1
aiofiles==23.2.1
